+++ src/main/java/roomescape/time/presentation/ReservationTimeController.java
@@ -1,19 +1,12 @@
 package roomescape.time.presentation;
 
-import jakarta.validation.Valid;
-import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.time.dto.AvailableTimeResponse;
-import roomescape.time.dto.ReservationTimeAddRequest;
 import roomescape.time.dto.ReservationTimeResponse;
 import roomescape.time.service.ReservationTimeService;
 
@@ -31,24 +24,10 @@ public class ReservationTimeController {
         return ResponseEntity.ok(reservationTimeService.findAllReservationTime());
     }
 
-    @PostMapping("/times")
-    public ResponseEntity<ReservationTimeResponse> saveReservationTime(
-            @Valid @RequestBody ReservationTimeAddRequest reservationTimeAddRequest) {
-        ReservationTimeResponse saveResponse = reservationTimeService.saveReservationTime(reservationTimeAddRequest);
-        URI createdUri = URI.create("/times/" + saveResponse.id());
-        return ResponseEntity.created(createdUri).body(saveResponse);
-    }
-
     @GetMapping("/times/available")
     public ResponseEntity<List<AvailableTimeResponse>> readTimesWithReservationStatus(
             @RequestParam(name = "date") LocalDate date,
             @RequestParam(name = "themeId") Long themeId) {
         return ResponseEntity.ok(reservationTimeService.findAllWithReservationStatus(date, themeId));
-    }
-
-    @DeleteMapping("/times/{id}")
-    public ResponseEntity<Void> removeReservationTime(@PathVariable("id") Long id) {
-        reservationTimeService.removeReservationTime(id);
-        return ResponseEntity.noContent().build();
     }
 }
